Rename modle folder and prepare the class to become an entity
