+++ src/modules/cars/entities/Category.ts
@@ -1,9 +1,19 @@
+import { Column, CreateDateColumn, Entity, PrimaryColumn } from "typeorm";
 import { v4 as uuidV4 } from "uuid";
 
+// @Entity({tableName})
+@Entity("categories")
 class Category {
+  @PrimaryColumn()
   id?: string;
+
+  @Column()
   name: string;
+
+  @Column()
   description: string;
+
+  @CreateDateColumn()
   created_at: Date;
 
   constructor() {
