+++ src/modules/cars/entities/Specification.ts
