+++ src/modules/cars/repositories/ICategoriesRepository.ts
@@ -1,4 +1,4 @@
-import { Category } from "../model/Category";
+import { Category } from "../entities/Category";
 import { ICreateCategoryDTO } from "./implementations/CategoriesRepository";
 
 interface ICategoriesRepository {
