+++ src/modules/cars/repositories/implementations/CategoriesRepository.ts
@@ -1,4 +1,4 @@
-import { Category } from "../../model/Category";
+import { Category } from "../../entities/Category";
 import { ICategoriesRepository } from "../ICategoriesRepository";
 
 interface ICreateCategoryDTO {
